+++ pkg/controller/kogitoinfra/errors.go
@@ -16,6 +16,7 @@ package kogitoinfra
 
 import (
 	"fmt"
+
 	"github.com/kiegroup/kogito-cloud-operator/pkg/apis/app/v1beta1"
 )
 
@@ -38,7 +39,7 @@ func (e reconciliationError) Error() string {
 func errorForResourceNotFound(kind, instance, namespace string) reconciliationError {
 	return reconciliationError{
 		Reason:     v1beta1.ResourceNotFound,
-		innerError: fmt.Errorf("%s instance(%s) not found in namespace %s", kind, instance, namespace),
+		innerError: fmt.Errorf("%s resource(%s) not found in namespace %s", kind, instance, namespace),
 	}
 }
 
@@ -56,6 +57,20 @@ func errorForUnsupportedAPI(instance *v1beta1.KogitoInfra) reconciliationError {
 			instance.Spec.Resource.APIVersion,
 			instance.Spec.Resource.Kind,
 			getSupportedResources()),
+	}
+}
+
+func errorForMissingResourceConfig(instance *v1beta1.KogitoInfra, configName string) reconciliationError {
+	return reconciliationError{
+		Reason:     v1beta1.ResourceMissingResourceConfig,
+		innerError: fmt.Errorf("Resource %s, configuration information (%s) is missing", instance.Name, configName),
+	}
+}
+
+func errorForResourceConfigError(instance *v1beta1.KogitoInfra, errorMsg string) reconciliationError {
+	return reconciliationError{
+		Reason:     v1beta1.ResourceConfigError,
+		innerError: fmt.Errorf("Error in configuration for the infrastructure resource %s. Error is: %s", instance.Name, errorMsg),
 	}
 }