+++ RELEASE_NOTES.md
@@ -6,6 +6,7 @@
 - [KOGITO-3937](https://issues.redhat.com/browse/KOGITO-3937) Move probe configuration in to `probes` subproperty
 - [KOGITO-3898](https://issues.redhat.com/browse/KOGITO-3898) Provide CLI flags for subset of probe parameters
 - [KOGITO-3968](https://issues.redhat.com/browse/KOGITO-3968) Add support to build and deploy sw.json and sw.yaml files from source
+- [KOGITO-2759](https://issues.redhat.com/browse/KOGITO-2759) Kogito operator support for MongoDB persistence
 
 ## Bug Fixes
 - [KOGITO-3866](https://issues.redhat.com/browse/KOGITO-3866) Operator print error logs when KogitoRuntime delete